backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
pydantic==2.8.2
pydantic-settings==2.4.0
redis==5.0.8
Jinja2==3.1.6
requests>=2.32.0